move most logic from raster driver to raster.py module
